requirements.txt
ipykernel
requests 
streamlit
pandas
numpy
matplotlib
seaborn
scikit-learn
plotly
xgboost
torch
torchvision
torchsummary
pandas
numpy
matplotlib
Pillow
tensorflow
statsmodels
python-dotenv
keras
openai
google-generativeai